# Feed cache: Use circular queue for zero-alloc of memory