--- a/disperser/dataapi/v2/feed_cache.go
+++ b/disperser/dataapi/v2/feed_cache.go
@@ -21,14 +21,13 @@ const (
 // cached segment, it'll extend the cache segment.
 // If there are more than maxItems in cache, it'll evict the oldest items.
 type FeedCache[T any] struct {
-	mu           sync.RWMutex
-	segment      *cacheEntry[T]
-	maxItems     int
-	fetchFromDB  func(ctx context.Context, start, end time.Time, order FetchOrder, limit int) ([]*T, error)
-	getTimestamp func(*T) time.Time
-
+	mu      sync.RWMutex
+	segment *CircularQueue[T]
 	// Async updates to the cache segment
 	updateWg *sync.WaitGroup
+
+	fetchFromDB  func(ctx context.Context, start, end time.Time, order FetchOrder, limit int) ([]*T, error)
+	getTimestamp func(*T) time.Time
 }
 
 func NewFeedCache[T any](
@@ -37,7 +36,7 @@ func NewFeedCache[T any](
 	timestampFn func(*T) time.Time,
 ) *FeedCache[T] {
 	return &FeedCache[T]{
-		maxItems:     maxItems,
+		segment:      NewCircularQueue[T](maxItems, timestampFn),
 		fetchFromDB:  fetchFn,
 		getTimestamp: timestampFn,
 		updateWg:     &sync.WaitGroup{},
@@ -49,18 +48,6 @@ func NewFeedCache[T any](
 type timeRange struct {
 	start time.Time
 	end   time.Time
-}
-
-// cacheEntry describes a segment of results fetched via fetchFromDB for a time range.
-//
-// It has the following properties:
-// - all data items that are in range [start, end) are in the segment
-// - no data items that are outside the range [start, end) are included in the segment
-//
-// The data items are in ascending order by timestamp.
-type cacheEntry[T any] struct {
-	timeRange
-	data []*T
 }
 
 // executionPlan describes the breakdown of a data fetch query [start, end) into sub ranges
@@ -138,7 +125,7 @@ func (c *FeedCache[T]) makePlan(start, end time.Time, queryOrder FetchOrder, lim
 	queryRange := timeRange{start: start, end: end}
 
 	// Handle no cache or no overlap cases together
-	if segment == nil || !queryRange.overlaps(segment.timeRange) {
+	if segment.size == 0 || !queryRange.overlaps(segment.timeRange) {
 		return executionPlan[T]{
 			// The data=nil, so it doesn't matter we fill the `before` or `after`
 			after: &queryRange,
@@ -146,7 +133,7 @@ func (c *FeedCache[T]) makePlan(start, end time.Time, queryOrder FetchOrder, lim
 	}
 
 	// Get cached data that's overlapping the query range
-	cachedOverlap := c.getCacheOverlap(segment.data, start, end, queryOrder, limit)
+	cachedOverlap := segment.QueryTimeRange(start, end, queryOrder, limit)
 
 	// The query range is fully contained in cache, it's a full cache hit
 	if !start.Before(segment.start) && !end.After(segment.end) {
@@ -307,6 +294,9 @@ func (c *FeedCache[T]) updateCache(result *executionResult[T]) {
 		return
 	}
 
+	c.mu.Lock()
+	defer c.mu.Unlock()
+
 	before, after := result.before, result.after
 	beforeData, afterData := result.beforeData, result.afterData
 
@@ -322,7 +312,7 @@ func (c *FeedCache[T]) updateCache(result *executionResult[T]) {
 				start = c.getTimestamp(beforeData[0])
 			}
 		}
-		c.mergeCache(beforeData, start, end)
+		c.segment.MergeTimeRange(beforeData, start, end)
 	}
 
 	if len(afterData) > 0 {
@@ -337,143 +327,8 @@ func (c *FeedCache[T]) updateCache(result *executionResult[T]) {
 				start = c.getTimestamp(afterData[0])
 			}
 		}
-		c.mergeCache(afterData, start, end)
+		c.segment.MergeTimeRange(afterData, start, end)
 	}
-}
-
-func (c *FeedCache[T]) mergeCache(data []*T, start, end time.Time) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	segment := c.segment
-
-	// No cache yet
-	if segment == nil {
-		c.setCacheSegment(data, start, end)
-		return
-	}
-
-	// No overlap with the cache
-	if !segment.overlaps(timeRange{start: start, end: end}) {
-		// Two special cases: non-overlapping but intervals are connected
-		if start.Equal(segment.end) {
-			c.setCacheSegment(append(segment.data, data...), segment.start, end)
-		}
-		if end.Equal(segment.start) {
-			c.setCacheSegment(append(data, segment.data...), start, segment.end)
-		}
-
-		// If it's a disconnected newer segment, it should replace existing cache
-		if start.After(segment.end) {
-			c.setCacheSegment(data, start, end)
-		}
-
-		return
-	}
-
-	// It's a sub range contained in existing cache, do nothing
-	if !start.Before(segment.start) && !end.After(segment.end) {
-		return
-	}
-
-	// The data is newer than cache, extend the cache
-	if end.After(segment.end) {
-		split := 0
-		for ; split < len(data); split++ {
-			if !c.getTimestamp(data[split]).Before(c.segment.end) {
-				break
-			}
-		}
-		if split < len(data) {
-			newData := append(segment.data, data[split:]...)
-			c.setCacheSegment(newData, segment.start, end)
-		}
-		return
-	}
-
-	// Now we must have start.Before(segment.start) && segment.start.Before(end)
-	split := len(data) - 1
-	for ; split >= 0; split-- {
-		if c.getTimestamp(data[split]).Before(segment.start) {
-			break
-		}
-	}
-	if split >= 0 {
-		newData := append(data[:split+1], segment.data...)
-		c.setCacheSegment(newData, start, segment.end)
-	}
-}
-
-func (c *FeedCache[T]) setCacheSegment(data []*T, start, end time.Time) {
-	// Ensure we don't exceed maxItems, prioritizing recent data
-	if len(data) > c.maxItems {
-		data = data[len(data)-c.maxItems:] // Keep newest
-		start = c.getTimestamp(data[0])
-	}
-	c.segment = &cacheEntry[T]{
-		timeRange: timeRange{
-			start: start,
-			end:   end,
-		},
-		data: data,
-	}
-}
-
-func (c *FeedCache[T]) getCacheOverlap(data []*T, start, end time.Time, queryOrder FetchOrder, limit int) []*T {
-	if len(data) == 0 {
-		return []*T{}
-	}
-
-	// Find start and end indices of the overlap
-	startIdx := -1
-	endIdx := -1
-
-	for i, item := range data {
-		timestamp := c.getTimestamp(item)
-
-		// Found the first item at or after the start time
-		if startIdx == -1 && !timestamp.Before(start) {
-			startIdx = i
-		}
-
-		// Found the first item at or past the end time (exclusive)
-		if !timestamp.Before(end) {
-			endIdx = i
-			break
-		}
-	}
-
-	// If we never found the end, set it to the end of data
-	if endIdx == -1 {
-		endIdx = len(data)
-	}
-
-	// No overlap found
-	if startIdx == -1 || startIdx >= endIdx {
-		return []*T{}
-	}
-
-	// Calculate how many items in the overlap
-	overlapCount := endIdx - startIdx
-
-	// Apply limit if needed
-	if limit > 0 && limit < overlapCount {
-		if queryOrder == Ascending {
-			// For ascending, take first 'limit' items
-			endIdx = startIdx + limit
-		} else {
-			// For descending, take last 'limit' items
-			startIdx = endIdx - limit
-		}
-	}
-
-	// Note: we need to make a copy of the overlap because the cache data can be mutated
-	// by other threads after this function returns (within this function, the caller
-	// makes sure a reader lock is held)
-	result := make([]*T, endIdx-startIdx)
-	copy(result, data[startIdx:endIdx])
-
-	return result
 }
 
 func reverseOrder[T any](data []*T) []*T {